+++ src/spectrochempy/lazyimport/api_methods.py
@@ -20,6 +20,9 @@ SpectroChemPy API.
 
 _LAZY_IMPORTS = {
     "CrossValidationResult": "spectrochempy.analysis.cross_validation",
+    "GroupKFold": "spectrochempy.analysis.model_selection",
+    "KFold": "spectrochempy.analysis.model_selection",
+    "LeaveOneOut": "spectrochempy.analysis.model_selection",
     "cross_validate": "spectrochempy.analysis.cross_validation",
     "PLSRegression": "spectrochempy.analysis.crossdecomposition.pls",
     "polynomial": "spectrochempy.analysis.curvefitting._models",